SDK/requirements.txt
pyserial==3.4
pip==20.1.1
pyserial==3.4
setuptools==47.1.0
pytest==6.2.2
pylint==2.6.0